Move all previews in commonMain (now compatible with new Kotlin Multiplatform plugin, except dark mode)
