+++ tasks-app-desktop/build.gradle.kts
@@ -160,7 +160,7 @@ compose.desktop {
 
 aboutLibraries {
     // - If the automatic registered android tasks are disabled, a similar thing can be achieved manually
-    // - `./gradlew :tasks-app-desktop:exportLibraryDefinitions -Pci=true`
+    // - `./gradlew :tasks-app-desktop:exportLibraryDefinitions --no-configuration-cache -Pci=true`
     // - the resulting file can for example be added as part of the SCM
     collect {
         configPath = file("$rootDir/license_config")
