--- tasks-app-shared/src/androidMain/kotlin/net/opatry/tasks/app/ui/component/ProfileIconUnsignedMenuPreview.kt
@@ -1,56 +1,0 @@
-/*
- * Copyright (c) 2025 Olivier Patry
- *
- * Permission is hereby granted, free of charge, to any person obtaining
- * a copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the Software
- * is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
- * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
- * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
- * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
- * OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package net.opatry.tasks.app.ui.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import net.opatry.tasks.app.presentation.UserState
-import net.opatry.tasks.app.ui.tooling.TaskfolioThemedPreview
-
-// FIXME using dark & light preview with popup doesn't work
-@Preview
-@Composable
-private fun ProfileIconUnsignedMenuPreview() {
-    TaskfolioThemedPreview(Modifier.size(width = 350.dp, height = 300.dp)) {
-        Column {
-            ProfileIcon(
-                userState = UserState.Unsigned,
-                showUserMenu = true,
-            ) {
-                Button(
-                    onClick = {},
-                    modifier = Modifier.align(Alignment.CenterHorizontally)
-                ) {
-                    Text("Authorize")
-                }
-            }
-        }
-    }
-}
