+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/component/TaskEditorBottomSheet.kt
@@ -45,16 +45,21 @@ import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExposedDropdownMenuBox
 import androidx.compose.material3.ExposedDropdownMenuDefaults
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.MenuAnchorType
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SheetState
+import androidx.compose.material3.SheetValue
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
+import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.rememberModalBottomSheetState
+import androidx.compose.material3.rememberStandardBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
@@ -68,6 +73,9 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.unit.dp
 import kotlinx.datetime.LocalDate
+import net.opatry.tasks.TodoTaskPosition
+import net.opatry.tasks.app.presentation.model.TaskId
+import net.opatry.tasks.app.presentation.model.TaskListId
 import net.opatry.tasks.app.presentation.model.TaskListUIModel
 import net.opatry.tasks.app.presentation.model.TaskUIModel
 import net.opatry.tasks.app.ui.component.TaskEditorBottomSheetTestTag.CANCEL_BUTTON
@@ -77,6 +85,7 @@ import net.opatry.tasks.app.ui.component.TaskEditorBottomSheetTestTag.SHEET_TITL
 import net.opatry.tasks.app.ui.component.TaskEditorBottomSheetTestTag.TITLE_FIELD
 import net.opatry.tasks.app.ui.component.TaskEditorBottomSheetTestTag.TITLE_FIELD_ERROR_MESSAGE
 import net.opatry.tasks.app.ui.component.TaskEditorBottomSheetTestTag.VALIDATE_BUTTON
+import net.opatry.tasks.app.ui.tooling.TaskfolioThemedPreview
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.dialog_cancel
 import net.opatry.tasks.resources.task_editor_sheet_edit_title
@@ -91,6 +100,9 @@ import net.opatry.tasks.resources.task_editor_sheet_title_field_label
 import net.opatry.tasks.resources.task_editor_sheet_title_field_placeholder
 import net.opatry.tasks.resources.task_editor_sheet_validate
 import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameter
+import org.jetbrains.compose.ui.tooling.preview.PreviewParameterProvider
 
 @VisibleForTesting
 object TaskEditorBottomSheetTestTag {
@@ -282,6 +294,98 @@ fun TaskEditorBottomSheet(
                 ) {
                     Text(stringResource(Res.string.task_editor_sheet_validate))
                 }
+            }
+        }
+    }
+}
+
+private data class TaskEditorPreviewData(
+    val editMode: TaskEditMode,
+    val task: TaskUIModel?,
+    val taskLists: List<TaskListUIModel> = emptyList(),
+    val taskList: TaskListUIModel = TaskListUIModel(
+        id = TaskListId(0),
+        title = "osef",
+    ),
+)
+
+private class TaskEditorPreviewParameterProvider : PreviewParameterProvider<TaskEditorPreviewData> {
+    override val values: Sequence<TaskEditorPreviewData>
+        get() = sequenceOf(
+            TaskEditorPreviewData(
+                TaskEditMode.NewTask,
+                task = null,
+                taskLists = emptyList(),
+                taskList = TaskListUIModel(
+                    id = TaskListId(0),
+                    title = "TODO",
+                ),
+            ),
+            TaskEditorPreviewData(
+                editMode = TaskEditMode.Edit,
+                task = TaskUIModel.Todo(
+                    id = TaskId(0),
+                    title = "My edited task",
+                    dueDate = LocalDate(2018, 6, 23),
+                    position = TodoTaskPosition.fromIndex(0),
+                ),
+                taskList = TaskListUIModel(
+                    id = TaskListId(0),
+                    title = "Current list",
+                ),
+            ),
+            TaskEditorPreviewData(
+                editMode = TaskEditMode.NewSubTask,
+                task = null,
+                taskList = TaskListUIModel(
+                    id = TaskListId(0),
+                    title = "Current list",
+                ),
+            )
+        )
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Preview
+@Composable
+private fun TaskEditorBottomSheetPreview(
+    @PreviewParameter(TaskEditorPreviewParameterProvider::class)
+    previewData: TaskEditorPreviewData
+) {
+    TaskfolioThemedPreview {
+        val sheetState = rememberStandardBottomSheetState(
+            initialValue = SheetValue.Expanded
+        )
+        Scaffold(
+            topBar = {
+                TopAppBar(title = { Text("Task Editor demo ${previewData.editMode.name}") })
+            }
+        ) { padding ->
+            Column(modifier = Modifier
+                .padding(padding)
+                .padding(horizontal = 24.dp)) {
+                Text("Some screen content", style = MaterialTheme.typography.titleLarge)
+                HorizontalDivider()
+                Text("Task lists", style = MaterialTheme.typography.titleMedium)
+                previewData.taskLists.forEach { taskList ->
+                    Text(taskList.title)
+                }
+                HorizontalDivider()
+                Text("Selected task list", style = MaterialTheme.typography.titleMedium)
+                Text(previewData.taskList.title)
+
+                // FIXME standard bottom sheet state hides the content
+
+                TaskEditorBottomSheet(
+                    editMode = previewData.editMode,
+                    task = previewData.task,
+                    allTaskLists = previewData.taskLists,
+                    sheetState = sheetState,
+                    selectedTaskList = previewData.taskList,
+                    onDismiss = {},
+                    onEditDueDate = {},
+                    onValidate = { _, _, _, _ -> },
+                )
             }
         }
     }
