+++ tasks-app-shared/src/commonMain/kotlin/net/opatry/tasks/app/ui/component/TaskListEditMenu.kt
@@ -22,35 +22,45 @@
 
 package net.opatry.tasks.app.ui.component
 
+import EllipsisVertical
 import LucideIcons
 import Trash2
 import androidx.annotation.VisibleForTesting
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.LocalContentColor
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
+import net.opatry.tasks.app.presentation.model.TaskListId
 import net.opatry.tasks.app.presentation.model.TaskListUIModel
 import net.opatry.tasks.app.ui.component.TaskListEditMenuTestTag.CLEAR_COMPLETED_TASKS
 import net.opatry.tasks.app.ui.component.TaskListEditMenuTestTag.DELETE
 import net.opatry.tasks.app.ui.component.TaskListEditMenuTestTag.DELETE_NOT_ALLOWED_NOTICE
 import net.opatry.tasks.app.ui.component.TaskListEditMenuTestTag.EDIT_MENU
 import net.opatry.tasks.app.ui.component.TaskListEditMenuTestTag.RENAME
+import net.opatry.tasks.app.ui.tooling.TaskfolioThemedPreview
 import net.opatry.tasks.resources.Res
 import net.opatry.tasks.resources.task_list_menu_clear_all_completed_tasks
 import net.opatry.tasks.resources.task_list_menu_default_list_cannot_be_deleted
 import net.opatry.tasks.resources.task_list_menu_delete
 import net.opatry.tasks.resources.task_list_menu_rename
 import org.jetbrains.compose.resources.stringResource
+import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @VisibleForTesting
 internal object TaskListEditMenuTestTag {
@@ -122,5 +132,31 @@ fun TaskListEditMenu(
             onClick = { onAction(TaskListEditMenuAction.Delete) },
             modifier = Modifier.testTag(DELETE),
         )
+    }
+}
+
+@Preview
+@Composable
+private fun TaskListMenuPreview() {
+    TaskfolioThemedPreview {
+        Box(
+            Modifier
+                .fillMaxWidth()
+                .height(400.dp)
+                .padding(24.dp), contentAlignment = Alignment.TopEnd
+        ) {
+            IconButton(onClick = {}) {
+                Icon(LucideIcons.EllipsisVertical, null)
+                TaskListEditMenu(
+                    taskList = TaskListUIModel(
+                        id = TaskListId(0L),
+                        title = "My task list",
+                    ),
+                    expanded = true,
+                    onDismiss = {},
+                    onAction = {},
+                )
+            }
+        }
     }
 }
